new app/requirements.txt
alright @ git+https://github.com/Kalebu/alright.git@main
fake-http-header~=0.1.2
selenium~=4.7.2
google-auth~=2.28.1
google-auth-oauthlib~=1.2.0
gspread~=5.7.2
pydantic~=2.3.0
fastapi~=0.109.1
pydantic-settings~=2.0.3
phonenumbers~=8.13.21
boto3~=1.28.57
botocore~=1.31.63
awscli~=1.29.63
alright~=2.61